% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_be_extract.R
\name{name_be_extract}
\alias{name_be_extract}
\title{name_be_extract}
\usage{
name_be_extract(
  data_to_detect,
  col_to_detect = NULL,
  reference = "all",
  numb_threshold = 500,
  length_threshold = 3,
  exclude_dico = NULL,
  spec_char_insens = FALSE,
  respect_boundaries = TRUE,
  nl_street = FALSE
)
}
\arguments{
\item{data_to_detect}{le data.frame d'interet}

\item{col_to_detect}{le nom de la colonne dans laquelle detecter les presnoms}

\item{reference}{le dictionnaire des prenoms de reference}

\item{numb_threshold}{le seuil de frequence minimale de prenoms pour le dictionnaire}

\item{length_threshold}{le seuil de longueur minimale de prenoms pour le dictionnaire}

\item{exclude_dico}{exclure les noms communs de la detection}

\item{spec_char_insens}{insensible aux caracteres speciaux dans la detection}

\item{respect_boundaries}{respecter les frontieres des mots}

\item{nl_street}{detection dans des noms de rue en neerlandais}
}
\description{
name_be_extract
}
\examples{
x <- data.frame(rue = c(paste0("Avenue Paul H","\u00e9","ger"),
"Jardin des Justes", "Place Sainte-Catherine"))

result <- name_be_extract(data_to_detect = x,
col_to_detect = "rue")
}
