% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_be_genderize.R
\name{name_be_genderize}
\alias{name_be_genderize}
\title{name_be_genderize}
\usage{
name_be_genderize(
  data_to_detect,
  col_name = NULL,
  reference = "all",
  inexact = FALSE,
  method_stringdist = "lcs",
  error_max = 1
)
}
\arguments{
\item{data_to_detect}{le data.frame d'interet}

\item{col_name}{le nom de la colonne a genderiser}

\item{reference}{le dictionnaire de prenoms de reference}

\item{inexact}{indiquer si le matching doit etre inexact}

\item{method_stringdist}{la methode de matching inexact}

\item{error_max}{l'erreur maximale du matching inexact}
}
\description{
name_be_genderize
}
\examples{
x <- data.frame(prenom = c(paste0("C","\u00e9","line"), "Najib", "Dominique"))

result <- name_be_genderize(data_to_detect = x,
col_name = "prenom")
}
