% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noms.R
\docType{data}
\name{noms}
\alias{noms}
\title{Donnees de prenoms en Belgique (2022)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 50883 rows and 9 columns.
}
\source{
\url{https://statbel.fgov.be/fr/themes/population/noms-et-prenoms/prenoms-femmes-et-hommes#figures}
}
\usage{
data(noms)
}
\description{
Les prenoms de toute la population en Belgique en 2022,
ainsi que les frequences de ces noms
}
\details{
#' @docType data
}
\keyword{datasets}
