% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noms_65.R
\docType{data}
\name{noms_65}
\alias{noms_65}
\title{Prenoms en Belgique des personnnes de 65 ans et plus (2009)}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6033 rows and 9 columns.
}
\source{
\url{https://statbel.fgov.be/fr/themes/population/noms-et-prenoms/prenoms-femmes-et-hommes#figures}
}
\usage{
data(noms_65)
}
\description{
Les prenoms pour la population des 65 ans et plus en Belgique en 2009,
ainsi que les frequences de ces noms
}
\details{
#' @docType data
}
\keyword{datasets}
